Finished roots for component data.
