+++ src/main/java/com/example/etl1/model/CpuCooler.java
@@ -1,4 +1,4 @@
-package com.example.etl1.Models;
+package com.example.etl1.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
@@ -6,31 +6,33 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Getter
 @Setter
 @NoArgsConstructor
 @Entity
-@Table(name = "POWER_SUPPLIES")
-public class PowerSupply {
+@Table(name = "CPU_COOLERS")
+public class CpuCooler {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
     private double price;
-    private String type;
-    private String efficiency;
-    private int wattage;
-    private String modular;
+    @OneToMany(mappedBy = "cpu_cooler")
+    private List<FanRpm> rpm;
+    @OneToMany(mappedBy = "cpu_cooler")
+    private List<FanNoiseLevel> noise_level;
     private String color;
+    private int size;
 
-    public PowerSupply(String name, double price, String type, String efficiency, int wattage, String modular, String color) {
+    public CpuCooler(String name, double price, List<FanRpm> rpm, List<FanNoiseLevel> noise_level, String color, int size) {
         this.name = name;
         this.price = price;
-        this.type = type;
-        this.efficiency = efficiency;
-        this.wattage = wattage;
-        this.modular = modular;
+        this.rpm = rpm;
+        this.noise_level = noise_level;
         this.color = color;
+        this.size = size;
     }
 }
