+++ src/main/java/com/example/etl1/model/FanRpm.java
@@ -1,4 +1,4 @@
-package com.example.etl1.Models;
+package com.example.etl1.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
@@ -11,17 +11,17 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @Entity
-@Table(name = "MEMORY_MODULES")
-public class MemoryModule {
+@Table(name = "FAN_RPM")
+public class FanRpm {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    private Integer module;
+    private Integer rpm;
     @ManyToOne
-    @JoinColumn(name = "memory_id")
-    private Memory memory;
+    @JoinColumn(name = "cpu_cooler_id")
+    private CpuCooler cpu_cooler;
 
-    public MemoryModule(Integer module) {
-        this.module = module;
+    public FanRpm(Integer rpm) {
+        this.rpm = rpm;
     }
 }
