+++ src/main/java/com/example/etl1/model/MemoryModule.java
@@ -1,4 +1,4 @@
-package com.example.etl1.Models;
+package com.example.etl1.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
@@ -11,17 +11,17 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @Entity
-@Table(name = "MEMORY_SPEED")
-public class MemorySpeed {
+@Table(name = "MEMORY_MODULES")
+public class MemoryModule {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    private Integer speed;
+    private Integer module;
     @ManyToOne
     @JoinColumn(name = "memory_id")
     private Memory memory;
 
-    public MemorySpeed(Integer speed) {
-        this.speed = speed;
+    public MemoryModule(Integer module) {
+        this.module = module;
     }
 }
