+++ src/main/java/com/example/etl1/model/PowerSupply.java
@@ -1,4 +1,4 @@
-package com.example.etl1.Models;
+package com.example.etl1.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
@@ -6,33 +6,31 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Getter
 @Setter
 @NoArgsConstructor
 @Entity
-@Table(name = "CPU_COOLERS")
-public class CpuCooler {
+@Table(name = "POWER_SUPPLIES")
+public class PowerSupply {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
     private double price;
-    @OneToMany(mappedBy = "cpu_cooler")
-    private List<FanRpm> rpm;
-    @OneToMany(mappedBy = "cpu_cooler")
-    private List<FanNoiseLevel> noise_level;
+    private String type;
+    private String efficiency;
+    private int wattage;
+    private String modular;
     private String color;
-    private int size;
 
-    public CpuCooler(String name, double price, List<FanRpm> rpm, List<FanNoiseLevel> noise_level, String color, int size) {
+    public PowerSupply(String name, double price, String type, String efficiency, int wattage, String modular, String color) {
         this.name = name;
         this.price = price;
-        this.rpm = rpm;
-        this.noise_level = noise_level;
+        this.type = type;
+        this.efficiency = efficiency;
+        this.wattage = wattage;
+        this.modular = modular;
         this.color = color;
-        this.size = size;
     }
 }
