+++ src/main/java/com/example/etl1/repository/CaseRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.Case;
+import com.example.etl1.model.Case;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CaseRepository extends CrudRepository<Case, Integer> { }
