+++ src/main/java/com/example/etl1/repository/CpuCoolerRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.CpuCooler;
+import com.example.etl1.model.CpuCooler;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CpuCoolerRepository extends CrudRepository<CpuCooler, Integer> { }
