+++ src/main/java/com/example/etl1/repository/CpuRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.Cpu;
+import com.example.etl1.model.Cpu;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CpuRepository extends CrudRepository<Cpu, Integer> { }
