+++ src/main/java/com/example/etl1/repository/FanNoiseLevelRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.FanNoiseLevel;
+import com.example.etl1.model.FanNoiseLevel;
 import org.springframework.data.repository.CrudRepository;
 
 public interface FanNoiseLevelRepository extends CrudRepository<FanNoiseLevel, Integer> { }
