+++ src/main/java/com/example/etl1/repository/GraphicsCardRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.GraphicsCard;
+import com.example.etl1.model.GraphicsCard;
 import org.springframework.data.repository.CrudRepository;
 
 public interface GraphicsCardRepository extends CrudRepository<GraphicsCard, Integer> { }
