+++ src/main/java/com/example/etl1/repository/InternalStorageRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.InternalStorage;
+import com.example.etl1.model.InternalStorage;
 import org.springframework.data.repository.CrudRepository;
 
 public interface InternalStorageRepository extends CrudRepository<InternalStorage, Integer> { }
