+++ src/main/java/com/example/etl1/repository/MemoryModuleRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.MemoryModule;
+import com.example.etl1.model.MemoryModule;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MemoryModuleRepository extends CrudRepository<MemoryModule, Integer> { }
