+++ src/main/java/com/example/etl1/repository/MemoryRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.Memory;
+import com.example.etl1.model.Memory;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MemoryRepository extends CrudRepository<Memory, Integer> { }
