+++ src/main/java/com/example/etl1/repository/MemorySpeedRepository.java
@@ -1,6 +1,6 @@
-package com.example.etl1.Repositories;
+package com.example.etl1.repository;
 
-import com.example.etl1.Models.MemorySpeed;
+import com.example.etl1.model.MemorySpeed;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MemorySpeedRepository extends CrudRepository<MemorySpeed, Integer> { }
