+++ src/main/resources/templates/components/cpu-coolers.html
@@ -2,7 +2,7 @@
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>CPUs</title>
+    <title>CPU Coolers</title>
 </head>
 <body>
 <ul>
